pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "barc4dip"
version = "2025.01.09"
description= "A Python package for digital image processing"
readme = "README.md"
license = "CECILL-2.1"
authors = [{ name = "Rafael Celestre", email = "rafael.celestre@esrf.eu"}]
keywords = ["digital image processing", "dip", "image quality"]
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
  "Intended Audience :: Science/Research",
  "Topic :: Scientific/Engineering :: Physics"
]
requires-python = ">=3.10"
dependencies = [
    "numpy",
    "scipy",
    "pillow",
    "jupyterlab",
    "ipympl"
]

[project.urls]
Homepage = "https://github.com/barc4/barc4dip"
Repository = "https://github.com/barc4/barc4dip"

[tool.setuptools.packages.find]
where = ["src"]